Add Connection class
This class is intended to hide most of the details of a database
connection. It only has four methods: connect, disconnect, execute and
executeforresult. The sql queries one provides to execute are currently
assumed to be in Sql Server syntax and is translated to the connection's
syntax by the execute method. This means that if one is using
parametrized sql, that sql should be rendered before the execute method
is used.