Remove required sql in examples/
The required sql found in the examples/ folder was taken directly from
CohortMethod, but that sql produces very sparse data which is sparser
than what HDPS is meant to use in its algorithm. These need to be
rewritten to fit into the HDPS paradigm.

In addidtion to this, there is now a check done by
prioritizeOptCovariates() to see if no optional covariates were
generated by the sql. The previous code would have resulted in an error
in that situation.