Add ICD-9 condition_era sql code
The previous condition_era code considered OMOP condition_concept_id
values, but this is more granular for the HDPS algorithm (as well as not
being hierarchical). The current code instead considers the equivalent
ICD-9 codes and groups on the first 3 digits. Currently the 3 digit
grouping is hard-coded, but that could presumably be changed to a
parameter.