Add getDbHdpsData() interface
The getDbHdpsData() function is added to provide a similar interface to
hdps as the getDbCohortData() function does for the Cohort Method
package. Currently the interface is implemented in mostly a stub fashion
so that certain functionality from Cohort Method also works for hdps.
However, as it stands, most functionality in Cohort Method is not
supported in hdps.