Update sql building cohorts
This now builds the same cohorts as Cohort Method for the cases I'm
currently interested in.

Warning: This will not agree with Cohort Method on many cases. The
current file has been stripped until it agrees with what I currently
need. It is a bit hacky and needs to be fixed in general, but for the
moment this serves my purposes.